use addurl to add the gbdmoz url files to gigablast. it should index just those dmoz urls, and not spider their links. it should ignore external errors like ETCPTIMEDOUT when indexing so it will be identical to dmoz.

+++ Msg1.cpp
@@ -358,8 +358,8 @@ bool Msg1::sendSomeOfList ( ) {
 	// little debug thing for genCatdb from msg9b's huge list add
 	if ( m_list->m_listSize > 10000000 )
 		log("msg1: adding chunk @ %li of %li bytes",
-		    dataStart - m_list->m_list ,
-		    m_list->m_listSize );
+		    (long)(dataStart - m_list->m_list) ,
+		    (long)m_list->m_listSize );
 
 	// . now send this list to the host
 	// . this returns false if blocked, true otherwise
